worker/requirements.txt
celery~=5.6.2
minio~=7.2.20
pydantic~=2.12.5
pydantic-settings~=2.13.1
pymupdf~=1.27.1
python-docx~=1.2.0
qdrant-client~=1.17.0
redis~=7.2.1
google-genai~=1.66.0
sqlalchemy~=2.0.48
paddleocr~=3.3.2
paddlepaddle~=3.2.2